Added devbox run and rm - removed auto update settings
